pyproject.toml
[tool.poetry]
name = "firebase-auth-admin-app"
version = "0.1.0"
description = "simple firebase auth admin app"
authors = ["dokxid <dan-du@live.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
firebase-admin = "^6.4.0"
python-dotenv = "^1.0.1"
rich = "^13.7.0"
click = "^8.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
